requirements.txt
python-dotenv
telethon
psycopg2-binary
sqlalchemy
dbt-postgres
pandas
ultralytics

